Create 3DPrinterTile
If you land on the 3D printer tile for the first time, you see it printing and have to move again. If you come back it's done and you can take the object it printed. If you take it you realize that it's an Enter key for a keyboard, otherwise you just leave the tile.